01_backend/requirements.txt
openai
fastapi
uvicorn
requests
ib-insync
python-dotenv
numpy
pandas
websockets
PyJWT
python-multipart
bcrypt
google-generativeai
nest-asyncio
thetadata
pytz
sqlalchemy
psycopg2-binary
